Completed first iteration of schedule header file, started schedule source file.

+++ SolOasis/ControlIntfc.h
@@ -7,8 +7,8 @@
 class ControlIntfc {
   public:
     virtual ~ControlIntfc(){};
-    virtual Status rotateHorizontal(int8_t dgr) = 0;
-    virtual Status rotateVertical(int8_t dgr) = 0;
+    virtual Status rotateHorizontal(int16_t dgr) = 0;
+    virtual Status rotateVertical(int16_t dgr) = 0;
 };
 
 #endif /* _CONTROL_INTFC_H_ */
