+++ SolOasis/ControlModule.cpp
@@ -1,5 +1,3 @@
-
-#include <stdlib.h>
 #include "ControlModule.h"
 
 ControlModule::ControlModule(){
@@ -10,11 +8,11 @@ ControlModule::~ControlModule(){
   
 }
 
-Status ControlModule::rotateHorizontal(int8_t dgr){
+Status ControlModule::rotateHorizontal(int16_t dgr){
   return OK;
 }
 
-Status ControlModule::rotateVertical(int8_t dgr){
+Status ControlModule::rotateVertical(int16_t dgr){
   return OK;
 }
